feat: toggle moved to a different file
