+++ src/js/categories.js
@@ -1,6 +1,6 @@
 import cards from '../../data/cards.js';
 // eslint-disable-next-line import/no-cycle
-import { gameToggle, checkToggle } from './script.js';
+import checkToggle, { gameToggle } from './toggle.js';
 // eslint-disable-next-line import/no-cycle
 import { resetGame } from './game.js';
 // eslint-disable-next-line import/no-cycle
